Cleanup of utility scripts
