+++ .gitignore
@@ -12,12 +12,11 @@
 /hydroshare/local_settings.py
 /hydroshare/static
 /log
-/nginx/config-files/hs-nginx.conf
 /nginx/.htpasswd
 /nginx/htpasswd
 /nginx/Dockerfile
+/nginx/config-files/hs-nginx.conf
 /nginx/config-files/nginx.conf-default
-/hydroshare/nginx_ip
 /static
 /tmp
 celeryd.pid
